:recycle: Move auth0 route to the relevant routes file
