+++ join_github_app/main/routes/join.py
@@ -1,19 +1,9 @@
-from flask import (
-    Blueprint,
-    current_app,
-    flash,
-    redirect,
-    render_template,
-    request,
-    session,
-)
-from join_github_app.main.routes.error import error
+from flask import (Blueprint, current_app, flash, redirect, render_template,
+                   request, session)
+
 from join_github_app.main.config.constants import ALLOWED_EMAIL_DOMAINS
 from join_github_app.main.middleware.auth import requires_auth
 from join_github_app.main.validators.index import is_valid_email_pattern
-from join_github_app.main.validators.join_github_form_auth0_user import (
-    JoinGithubFormAuth0User,
-)
 
 join_route = Blueprint('join_route', __name__)
 
@@ -103,28 +93,3 @@ def is_digital_justice_email(domain):
 
 def is_justice_email(domain):
     return "justice.gov.uk" == domain.lower()
-
-
-def _join_github_auth0_users(request):
-    form = JoinGithubFormAuth0User(request.form)
-    if request.method == "POST" and form.validate() and form.validate_org():
-        selected_orgs = current_app.github_script.get_selected_organisations(
-            form.access_moj_org.data, form.access_as_org.data
-        )
-
-        if current_app.github_script.is_github_seat_protection_enabled() is True:
-            return error("GitHub Seat protection enabled")
-        else:
-            user_email = session["user"]["userinfo"]["email"]
-            current_app.github_script.add_new_user_to_github_org(
-                user_email, selected_orgs
-            )
-
-        return redirect("thank-you")
-
-    # Problem in the form
-    return render_template(
-        "pages/join-github-auth0-user.html",
-        form=form,
-        template="join-github-auth0-user.html",
-    )
